% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gauss_kruger.R
\name{check_gauss_kruger_3_deg}
\alias{check_gauss_kruger_3_deg}
\title{Calculate the 3-degree Gauss Kruger zone number and the central longitude of the input longitude range}
\usage{
check_gauss_kruger_3_deg(lonmax, lonmin)
}
\arguments{
\item{lonmax}{A numerical value, max longitude of the map extent}

\item{lonmin}{A numerical value, min longitude of the map extent}
}
\value{
A list with name of "zone_num" and "mid_lon", if the input range doesn't fall in any zone, the list will be NAs.
}
\description{
Calculate the 3-degree Gauss Kruger zone number and the central longitude of the input longitude range
}
\examples{
check_gauss_kruger_3_deg(11,12)
}
