% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_utm_zone.R
\name{check_utm_zone}
\alias{check_utm_zone}
\title{Calculate the UTM zone number and the proj4 string}
\usage{
check_utm_zone(obj)
}
\arguments{
\item{obj}{Input geo data, should be one of:\cr
\itemize{
\item An object of class \code{sf} / \code{sfc} / \code{bbox} (from \code{\link[sf:st_bbox]{sf::st_bbox()}})  \cr
\item A named list with longitude and latitude extents with names of "xmin", "xmax" "ymin" and "ymax"
}}
}
\value{
A list with "utm_zone_num", "hemisphere" and "proj4", if the boundary you input doensn't fall in any of the utm zone, all these values would be NA
}
\description{
Calculate the UTM zone number and the proj4 string
}
\examples{
check_utm_zone(c(xmin = 6.15, xmax = 6.3, ymin = 49.7, ymax = 49.8))
}
