% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConicOK.R
\name{checkConicOK}
\alias{checkConicOK}
\title{Title}
\usage{
checkConicOK(lat0, lon0, projectionString, latmax, latmin, lonmax, lonmin)
}
\arguments{
\item{lonmin}{}
}
\description{
Title
}
\keyword{internal}
