% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_country}
\alias{example_country}
\title{1:110m Spain & Bolivia map}
\format{
\subsection{"example_country"}{

A list with Spain and Bolivia 110m map from Natural Earth 1:110m Cultural Vectors
}
}
\source{
\url{https://www.naturalearthdata.com/downloads/110m-cultural-vectors/}
}
\usage{
example_country
}
\description{
1:110m Spain & Bolivia map
}
\examples{
spain_polygon = sf::st_polygon(list(projWiz::example_country$Spain))
plot(sf::st_sfc(spain_polygon, crs = 4326), graticule = TRUE, key.pos = NULL, axes = TRUE)

}
\keyword{datasets}
