% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConicOK.R
\name{normalizeLON}
\alias{normalizeLON}
\title{Title Longitude correction}
\usage{
normalizeLON(lon, lon0)
}
\arguments{
\item{lon}{Longitude value to be corrected}

\item{lon0}{Median longitude value}
}
\value{
A numerical value of corrected longitude value
}
\description{
Title Longitude correction
}
\examples{
normalizeLON()
}
\keyword{internal}
