% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEWextent.R
\name{printEWextent}
\alias{printEWextent}
\title{Funcion for regional maps with an east-west extent}
\usage{
printEWextent(property, center, latmax, latmin, lonmax, lonmin)
}
\arguments{
\item{property}{Longitude value to be corrected}

\item{center}{Median longitude value}

\item{latmax}{description}

\item{latmin}{description}

\item{lonmax}{description}

\item{lonmin}{description}
}
\value{
A string
}
\description{
Funcion for regional maps with an east-west extent
}
\examples{
printEWextent("Equalarea", center, latmax, latmin)
}
\keyword{internal}
