% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_equal_area.R
\name{proj_equal_area}
\alias{proj_equal_area}
\title{Projection for the area of interest with equal area projection}
\usage{
proj_equal_area(obj, output_type = "proj4")
}
\arguments{
\item{obj}{An object to compute the bounding box from, which can be accepted by \code{\link[sf:st_bbox]{sf::st_bbox()}}}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}
}
\value{
A \code{proj4} or \code{WKT} string
}
\description{
Auto selecting equal-area projections based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected.
}
\examples{
proj_equal_area(spData::alaska)
}
