% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_hemisphere.R
\name{proj_hemisphere}
\alias{proj_hemisphere}
\title{Projection for hemishpere}
\usage{
proj_hemisphere(
  obj,
  property = "ortho",
  output_type = "proj4",
  datum = "WGS84",
  unit = "m"
)
}
\arguments{
\item{obj}{An object to compute the bounding box from, which can be accepted by \code{\link[sf:st_bbox]{sf::st_bbox()}}}

\item{property}{Projection property, should be one of "Equalarea", "Conformal" and "ortho", the default value is ortho, for orthographic projection.}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}

\item{datum}{A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)}

\item{unit}{A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)}
}
\value{
A \code{proj4} or \code{WKT} string
}
\description{
Projection for hemishpere
}
\examples{
proj_hemisphere(sf::st_as_sf(quakes,coords = c("long","lat"),crs = 4326))
}
