% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_region.R
\name{proj_region}
\alias{proj_region}
\title{Projection for the area of interest with specified way of projection}
\usage{
proj_region(obj, property, output_type = "proj4")
}
\arguments{
\item{obj}{An object to compute the bounding box from, which can be accepted by \code{\link[sf:st_bbox]{sf::st_bbox()}}}

\item{property}{Projection property, should be one of "Equalarea", "Conformal" and "Equidistant".}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}
}
\value{
A \code{proj4} or \code{WKT} string
}
\description{
Auto selecting projections with specified property based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected.
}
\examples{
proj_conformal(spData::alaska, "Equalarea")
}
