% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_specify.R
\name{proj_specify}
\alias{proj_specify}
\title{Generating projection by the geo-boundary and the shorthand name}
\usage{
proj_specify(
  obj,
  prj,
  output_type = "proj4",
  datum = "WGS84",
  unit = "m",
  x0 = NA_real_
)
}
\arguments{
\item{obj}{An object to compute the bounding box from, which can be accepted by \code{\link[sf:st_bbox]{sf::st_bbox()}}}

\item{prj}{A string representing the shorthand name of the desired projection}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}

\item{datum}{A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)}

\item{unit}{A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)}

\item{x0}{A numeric value specifying the longitude offset}
}
\value{
A \code{proj4} or \code{WKT} string as \code{output_type} specified
}
\description{
Generating projection by the geo-boundary and the shorthand name
}
\examples{
proj_specify(sf::st_bbox(c(xmin = -30,ymin = -25,xmax = 20,ymax = 5),crs = 4326), "lcc")
}
