% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_world.R
\name{proj_world}
\alias{proj_world}
\title{Specify a world map projection}
\usage{
proj_world(world_proj, lng = 0, output_type = "proj4")
}
\arguments{
\item{world_proj}{A string of proj shorthand, could be extracted from \code{world_proj_list}}

\item{lng}{A numerical value of the expected central longitude}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}
}
\value{
A \code{proj4} or \code{WKT} string
}
\description{
Specify a world map projection
}
\examples{
proj_world(world_proj_list$compromise$barrel_shape$Natural_Earth, 150)
}
