% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utm_zone}
\alias{utm_zone}
\title{UTM zone geometry data}
\format{
\subsection{"utm_zone"}{

A Simple feature collection with 120 rows and 3 columns:
\describe{
\item{zone}{UTM zone number}
\item{hemisphere}{"N" for northern hemisphere, "S" for southern hemisphere}
\item{geometry}{UTM zone geometries}
}
}
}
\source{
\url{https://hub.arcgis.com/datasets/esri::world-utm-grid}
}
\usage{
utm_zone
}
\description{
UTM zone geometry data
}
\keyword{internal}
