% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_shaded_relief}
\alias{world_shaded_relief}
\title{Downsampled Natural Earth 1:50 Shaded Relief World Map}
\format{
\subsection{"world_shaded_relief"}{

A array of dimention of 360, 720, 3 for 3 layer for R, G, B band with 0.5 degree on lon and lat direction, please use \code{terra::rast()} when using
}
}
\source{
\url{https://www.naturalearthdata.com/downloads/50m-raster-data/50m-cross-blend-hypso/}
}
\usage{
world_shaded_relief
}
\description{
Downsampled Natural Earth 1:50 Shaded Relief World Map
}
\examples{
world_base_map = terra::rast(projWiz::world_shaded_relief)
terra::ext(world_base_map) <- c(-180,180,-90,90)

}
\keyword{internal}
