% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConicOK.R
\name{checkConicOK}
\alias{checkConicOK}
\title{Check if Conic projection will have cone at polars}
\usage{
checkConicOK(lat0, lon0, projectionString, latmax, latmin, lonmax, lonmin)
}
\arguments{
\item{lat0}{A numerical value, central latitude of the map extent}

\item{lon0}{A numerical value, central longitude of the map extent}

\item{projectionString}{A string, the full name of the projection, should be one of "Lambert conformal conic", "Albers equal area conic" and "Lambert azimuthal equal area"}

\item{latmax}{A numerical value, max latitude of the map extent}

\item{latmin}{A numerical value, min latitude of the map extent}

\item{lonmax}{A numerical value, max longitude of the map extent}

\item{lonmin}{A numerical value, min longitude of the map extent}
}
\value{
A integer value from -1/0/1, 1 means passing the test, -1 means failing to pass the test, 0 means the case of Albers when the fan of the selected extent spans less than 180° around a pole
}
\description{
Check if Conic projection will have cone at polars
}
\examples{
checkConicOK(48.5, 140,"Lambert azimuthal equal area",55,42,-132,52)
}
\keyword{internal}
