% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEWextent.R
\name{printEWextent}
\alias{printEWextent}
\title{Funcion for regional maps with an east-west extent}
\usage{
printEWextent(property, center, latmax, latmin, dlon, datum, unit)
}
\arguments{
\item{property}{Projection property, should be one of "Equalarea", "Conformal" and "Equidistant".}

\item{center}{A list with with numerical values named with \code{lng} & \code{lat} element, which are the median longitude and latitude values}

\item{latmax}{A numerical value, max latitude of the map extent}

\item{latmin}{A numerical value, min latitude of the map extent}

\item{dlon}{A numerical value, Longitudal extent of the region}

\item{datum}{A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)}

\item{unit}{A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)}
}
\value{
A list of strings named with \code{PROJ} & \code{WKT}
}
\description{
Funcion for regional maps with an east-west extent
}
\keyword{internal}
