% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSquareFormat.R
\name{printSquareFormat}
\alias{printSquareFormat}
\title{Funcion for regional maps with an square-shaped extent}
\usage{
printSquareFormat(property, center, latmax, latmin)
}
\arguments{
\item{property}{Projection property, should be one of "Equalarea", "Conformal" and "Equidistant".}

\item{center}{A list with with numerical values named with \code{lng} & \code{lat} element, which are the median longitude and latitude values}

\item{latmax}{A numerical value, max latitude of the map extent}

\item{latmin}{A numerical value, min latitude of the map extent}
}
\value{
A list of strings named with \code{PROJ} & \code{WKT}
}
\description{
Funcion for regional maps with an square-shaped extent
}
\examples{
printSquareFormat("Equalarea", center, latmax, latmin)
}
\keyword{internal}
