% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_equal_area.R
\name{proj_equal_area}
\alias{proj_equal_area}
\title{Projection for the area of interest with equal area projection}
\usage{
proj_equal_area(obj, output_type = "proj4", datum = "WGS84", unit = "m")
}
\arguments{
\item{obj}{Input geo data, should be one of:\cr
\itemize{
\item An object can be accepted by \code{\link[sf:st_bbox]{sf::st_bbox()}} to compute the bounding box\cr
\item A named list with longitude and latitude extents with names of "xmin", "xmax" "ymin" and "ymax"
}}

\item{output_type}{A string for expected output, either "proj4" or "WKT"}

\item{datum}{A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)}

\item{unit}{A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)}
}
\value{
A \code{proj4} or \code{WKT} string
}
\description{
Auto selecting equal-area projections based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected.\cr
Please note that there's a longitudinal (less than 160) and latitudinal (less than 80) range limit. If area of interest larger than this please use \code{\link[=proj_hemisphere]{proj_hemisphere()}} or use \code{\link[=proj_specify]{proj_specify()}}.
}
\examples{
proj_equal_area(c(xmax=112,xmin=156,ymin=6,ymax=23))
}
\seealso{
\code{\link[=proj_region]{proj_region()}}, \code{\link[=proj_conformal]{proj_conformal()}}, \code{\link[=proj_equidistant]{proj_equidistant()}}
}
