% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringLinks.R
\name{stringLinks}
\alias{stringLinks}
\title{Function to generate proj4 and WKT string}
\usage{
stringLinks(prj, x0, lat0, lat1, lat2, lon0, k0)
}
\arguments{
\item{prj}{A string of the shorthand of desired prjection}

\item{x0}{A numerical value for offset}

\item{lat0}{A numerical value for centeral latitude}

\item{lat1}{A numerical value for standard parallel 1}

\item{lat2}{A numerical value for standard parallel 2}

\item{lon0}{A numerical value for centeral longitude}

\item{k0}{A numerical value for scale factor}
}
\value{
A list of strings named with \code{PROJ} & \code{WKT}
}
\description{
Function to generate proj4 and WKT string
}
\examples{
stringLinks("aeqd", NaN, 90.0, NaN, NaN, 130, NaN)
}
\keyword{internal}
