% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringLinks.R
\name{stringLinks}
\alias{stringLinks}
\title{Function to generate proj4 and WKT string}
\usage{
stringLinks(prj, x0, lat0, lat1, lat2, lon0, k0, datum = "WGS84", unit = "m")
}
\arguments{
\item{prj}{A string representing the shorthand name of the desired projection}

\item{x0}{A numeric value specifying the longitude offset}

\item{lat0}{A numerical value for centeral latitude}

\item{lat1}{A numerical value for standard parallel 1}

\item{lat2}{A numerical value for standard parallel 2}

\item{lon0}{A numerical value for centeral longitude}

\item{k0}{A numerical value for scale factor}

\item{datum}{A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)}

\item{unit}{A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)}
}
\value{
A list of strings named with \code{PROJ} & \code{WKT}
}
\description{
Function to generate proj4 and WKT string
}
\keyword{internal}
