% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_proj_list}
\alias{world_proj_list}
\title{World Map projection list}
\format{
A list with popular world projections:
\describe{
\item{equal_area}{Projections to keep the area approximately correct}
\item{point_polar}{Equal-area world map projections with poles represented as points}
\item{line_polar}{Equal-area world map projections with poles represented as lines}
\item{compromise}{Projections compromising shape and area}
\item{round_boudnary}{Boundary on east and west as curve}
\item{rectangular}{Boundary as rectange}
}
}
\source{
\url{https://github.com/ProjectionWizard/projectionwizard.github.io/blob/263b9ff09128e371ee923ab57bfc1ed41bbdc4ba/outputFormat.js#L115}
}
\usage{
world_proj_list
}
\description{
World Map projection list
}
\keyword{datasets}
